# docs/swagger.json
{
    "openapi": "3.1.0",
    "info": {
      "version": "1.0.0",
      "title": "Simple contacts App",
      "license": {
        "name": "Apache 2.0",
        "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
      },
      "description": "Very simple minimalistic contacts app."
    },
    "servers": [
      {
        "url": "http://localhost:3000"
      },
      {
        "url": "https://hw2-mongodb-41jp.onrender.com"
      }
    ],
...
    "paths": {
...
      "/auth/login": {
        "post": {
          "tags": [
            "Auth"
          ],
          "summary": "Login user",
          "operationId": "loginUser",
          "description": "Logs in a user with email and password.",
          "security": [],
          "requestBody": {
            "description": "Credentials for login.",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "email",
                    "password"
                  ],
                  "properties": {
                    "email": {
                      "type": "string",
                      "format": "email",
                      "description": "User email address.",
                      "example": "jconnor@skynet.mil"
                    },
                    "password": {
                      "type": "string",
                      "format": "password",
                      "description": "User password.",
                      "example": "SecretPassword"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successfully user logged.",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "required": [
                      "status",
                      "message",
                      "data"
                    ],
                    "properties": {
                      "status": {
                        "type": "integer",
                        "example": 200
                      },
                      "message": {
                        "type": "string",
                        "example": "Successfully logged in an user!"
                      },
                      "data": {
                        "type": "object",
                        "properties": {
                          "accessToken": {
                            "type": "string",
                            "description": "Access token.",
                            "example": "nWlVaC8bxPl9psLO8sCM0MFjFUbcxnXXOPgoPNwj"
                          }
                        }
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "$ref": "#/components/responses/400"
            },
            "401": {
              "$ref": "#/components/responses/401"
            },
            "404": {
              "$ref": "#/components/responses/404"
            },
            "500": {
              "$ref": "#/components/responses/500"
            }
          }
        }
      },
      "/auth/logout": {
        "post": {
          "tags": [
            "Auth"
          ],
          "summary": "User logout",
          "operationId": "logoutUser",
          "description": "User log out.",
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "parameters": [
            {
              "in": "cookie",
              "name": "sessionId",
              "description": "User session Id.",
              "required": true,
              "schema": {
                "type": "string",
                "example": "j%3A%2267967ae6737a46114f32bb6a%22"
              }
            }
          ],
          "responses": {
            "204": {
              "description": "Successfully user logout. No content returned."
            },
            "400": {
              "$ref": "#/components/responses/400"
            },
            "500": {
              "$ref": "#/components/responses/500"
            }
          }
        }
      },
...
      "/auth/reset-pwd": {
        "post": {
          "tags": [
            "Auth"
          ],
          "summary": "Reset password",
          "operationId": "resetPassword",
          "description": "Reset password with reset token and new password.",
          "security": [],
          "requestBody": {
            "description": "Reset token from email and new password.",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "token",
                    "password"
                  ],
                  "properties": {
                    "token": {
                      "type": "string",
                      "description": "Reset token.",
                      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI2NzhiZjgzZTE0ZDU1OTg1ZjA5MDBmNjAiLCJlbWFpbCI6ImR2YW5rZXZpY2hANXByby1zb2Z0d2FyZS5uZXQiLCJpYXQiOjE3MzcyMjkwOTMsImV4cCI6MTczNzIyOTk5M30.Eloifgrr6wRLAxLXaMunnbjlaR7QPrU89JVcp2i0"
                    },
                    "password": {
                      "type": "string",
                      "format": "password",
                      "description": "New user password.",
                      "example": "NewPa$$word"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successfully updated password.",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "required": [
                      "status",
                      "message",
                      "data"
                    ],
                    "properties": {
                      "status": {
                        "type": "integer",
                        "example": 200
                      },
                      "message": {
                        "type": "string",
                        "example": "Password was successfully reset!"
                      },
                      "data": {
                        "type": "object",
                        "example": {}
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "$ref": "#/components/responses/400"
            },
            "401": {
              "$ref": "#/components/responses/401"
            },
            "404": {
              "$ref": "#/components/responses/404"
            },
            "500": {
              "$ref": "#/components/responses/500"
            }
          }
        }
      }
    },
    "components": {
      "securitySchemes": {
        "bearerAuth": {
          "type": "http",
          "scheme": "bearer"
        }
      },
      "responses": {
        "400": {
          "description": "Bad Request",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status",
                  "message",
                  "data"
                ],
                "properties": {
                  "status": {
                    "type": "integer",
                    "example": 400
                  },
                  "message": {
                    "type": "string",
                    "example": "BadRequestError"
                  },
                  "data": {
                    "type": "object",
                    "required": [
                      "message"
                    ],
                    "properties": {
                      "message": {
                        "type": "string",
                        "example": "Bad Request. Invalid parameter."
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "401": {
          "description": "Unauthorized error",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status",
                  "message",
                  "data"
                ],
                "properties": {
                  "status": {
                    "type": "integer",
                    "example": 401
                  },
                  "message": {
                    "type": "string",
                    "example": "UnauthorizedError"
                  },
                  "data": {
                    "type": "object",
                    "required": [
                      "message"
                    ],
                    "properties": {
                      "message": {
                        "type": "string",
                        "example": "Access token expired or Session not found."
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "404": {
          "description": "Not Fount Error",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status",
                  "message",
                  "data"
                ],
                "properties": {
                  "status": {
                    "type": "integer",
                    "example": 404
                  },
                  "message": {
                    "type": "string",
                    "example": "NotFoundError"
                  },
                  "data": {
                    "type": "object",
                    "required": [
                      "message"
                    ],
                    "properties": {
                      "message": {
                        "type": "string",
                        "example": "Contact not found"
                      }
                    }
                  }
                }
              }
            }
          }
        },
...
        "500": {
          "description": "Internal Server Error",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status",
                  "message",
                  "data"
                ],
                "properties": {
                  "status": {
                    "type": "integer",
                    "example": 500
                  },
                  "message": {
                    "type": "string",
                    "example": "Internal Server Error"
                  },
                  "data": {
                    "type": "object",
                    "required": [
                      "message"
                    ],
                    "properties": {
                      "message": {
                        "type": "string",
                        "example": "Something went wrong"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
...
    }
  }